# documentation API/G-MOOC 4D (Admin).postman_collection.json-SwaggerYaml.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: G-MOOC 4D (Admin)
  contact: {}
host: nurz.me
basePath: /
securityDefinitions: {}
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/admin/materi/update/1:
    put:
      summary: update materi
      tags:
        - Materi
      operationId: updatemateri
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: Bearer {token}
          type: string
        - name: Body
          in: body
          required: true
          description: ""
          schema:
            $ref: "#/definitions/updatematerirequest"
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/updatemateri"
          examples:
            application/json:
              metadata:
                code: 201
                status: success
                message: Materi Berhasil Diperbarui
              data:
                name: Javascrypt
                id_kelas: 1
          headers:
            Host:
              type: string
              default: localhost:8000
            Date:
              type: string
              default: Wed, 09 Aug 2023 04:13:00 GMT
            Connection:
              type: string
              default: close
            X-Powered-By:
              type: string
              default: PHP/8.2.8
            Cache-Control:
              type: string
              default: no-cache, private
            X-RateLimit-Limit:
              type: string
              default: "60"
            X-RateLimit-Remaining:
              type: string
              default: "59"
            Vary:
              type: string
              default: Origin
...
  /api/admin/quiz/update/1:
    put:
      summary: Update Quiz
      tags:
        - Quiz
      operationId: UpdateQuiz
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: false
          default: Bearer {token}
          type: string
        - name: Body
          in: body
          required: true
          description: ""
          schema:
            $ref: "#/definitions/UpdateQuizRequest"
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/UpdateQuiz"
          examples:
            application/json:
              metadata:
                code: 201
                status: success
                message: Quiz berhasil diubah
              data:
                id_kelas: 1
                question: Bagaimana cara memberikan ini
                true_answer: A
                option_A: ini adalah a
                option_B: ini adalah b
                option_C: ini adalah c
          headers:
            Host:
              type: string
              default: localhost:8000
            Date:
              type: string
              default: Fri, 11 Aug 2023 11:56:37 GMT
            Connection:
              type: string
              default: close
            X-Powered-By:
              type: string
              default: PHP/8.1.12-1ubuntu4.2
            Cache-Control:
              type: string
              default: no-cache, private
            X-RateLimit-Limit:
              type: string
              default: "60"
            X-RateLimit-Remaining:
              type: string
              default: "58"
            Vary:
              type: string
              default: Origin
...
definitions:
  Metadata:
    title: Metadata
    example:
      code: 200
      status: success
      message: User Berhasil Masuk
    type: object
    properties:
      code:
        type: integer
        format: int32
      status:
        type: string
      message:
        type: string
    required:
      - code
      - status
      - message
...
  updatematerirequest:
    title: Update Materi Request
    example:
      id_kelas: 1
      name: Pengenalan Javascrypt
      materi: JavaScript adalah suatu bahasa pemrograman tingkat tinggi dan dinamis. JavaScript populer di internet dan dapat bekerja di sebagian besar penjelajah web populer seperti Google Chrome, Internet Explorer, Mozilla Firefox, Netscape dan Opera. Kode JavaScript dapat disisipkan dalam halaman web menggunakan tag script.
      url: https://www.youtube.com/embed/upDLs1sn7g4
      durasi: 54
      poin: 20
    type: object
    properties:
      id_kelas:
        type: integer
        format: int32
      name:
        type: string
      materi:
        type: string
      url:
        type: string
      durasi:
        type: integer
        format: int32
      poin:
        type: integer
        format: int32
    required:
      - id_kelas
      - name
      - materi
      - url
      - durasi
      - poin
  updatemateri:
    title: Update materi
    example:
      metadata:
        code: 201
        status: success
        message: Materi Berhasil Diperbarui
      data:
        name: Javascrypt
        id_kelas: 1
    type: object
    properties:
      metadata:
        $ref: "#/definitions/Metadata"
      data:
        $ref: "#/definitions/DataUpdateMateri"
    required:
      - metadata
      - data
  DataUpdateMateri:
    title: Data update materi
    example:
      name: Javascrypt
      id_kelas: 1
    type: object
    properties:
      name:
        type: string
      id_kelas:
        type: integer
        format: int32
    required:
      - name
      - id_kelas
...
  DataCreateQuiz:
    title: Data Create Quiz
    example:
      id_kelas: 1
      question: Bagaimana cara memberikan ini
      true_answer: A
      option_A: ini adalah a
      option_B: ini adalah b
      option_C: ini adalah c
    type: object
    properties:
      id_kelas:
        type: integer
        format: int32
      question:
        type: string
      true_answer:
        type: string
      option_A:
        type: string
      option_B:
        type: string
      option_C:
        type: string
    required:
      - id_kelas
      - question
      - true_answer
      - option_A
      - option_B
      - option_C
...
  UpdateQuizRequest:
    title: Update quiz request
    example:
      id_kelas: 2
      question: Pengertian dari Javascript
    type: object
    properties:
      id_kelas:
        type: integer
        format: int32
      question:
        type: string
    required:
      - id_kelas
      - question
  UpdateQuiz:
    title: Update quiz
    example:
      metadata:
        code: 201
        status: success
        message: Quiz berhasil diubah
      data:
        id_kelas: 1
        question: Bagaimana cara memberikan ini
        true_answer: A
        option_A: ini adalah a
        option_B: ini adalah b
        option_C: ini adalah c
    type: object
    properties:
      metadata:
        $ref: "#/definitions/Metadata"
      data:
        $ref: "#/definitions/DataCreateQuiz"
    required:
      - metadata
      - data
...
security: []
...